Refactor Login routing to use view instead of directly routing to form component
