+++ src/views/LoginPage.vue
@@ -1,0 +1,9 @@
+<script setup>
+  import FormLogin from '@/components/FormLogin.vue'
+</script>
+
+<template>
+  <div class="login-page">
+    <FormLogin />
+  </div>
+</template>
